fixed some win32 warnings
git-svn-id: https://btanks.svn.sourceforge.net/svnroot/btanks/trunk@4520 5c0bf991-d514-0410-a65f-ce4832e571f3

+++ src/special_zone.cpp
@@ -68,7 +68,7 @@ const bool SpecialZone::final() const {
 }
 
 void SpecialZone::onTimer(const int slot_id, const bool win) {
-	float duration = atof(subname.c_str());
+	float duration = (float)atof(subname.c_str());
 	LOG_DEBUG(("activating timer %s for %g seconds", name.c_str(), duration));
 	PlayerSlot &slot = PlayerManager->getSlot(slot_id);
